+++ cpp/src/IceUtil/Time.cpp
@@ -144,7 +144,7 @@ IceUtil::Time::now(Clock clock)
         }
         return Time(tv.tv_sec * ICE_INT64(1000000) + tv.tv_usec);
 #elif defined(__APPLE__)
-        return Time(static_cast<IceUtil::Int64>(mach_absolute_time() * machMultiplier));
+        return Time(static_cast<IceUtil::Int64>(static_cast<double>(mach_absolute_time()) * machMultiplier));
 #else
         struct timespec ts;
         if(clock_gettime(CLOCK_MONOTONIC, &ts) < 0)
@@ -224,13 +224,13 @@ IceUtil::Time::toMicroSeconds() const
 double
 IceUtil::Time::toSecondsDouble() const
 {
-    return _usec / 1000000.0;
+    return static_cast<double>(_usec) / 1000000.0;
 }
 
 double
 IceUtil::Time::toMilliSecondsDouble() const
 {
-    return _usec / 1000.0;
+    return static_cast<double>(_usec) / 1000.0;
 }
 
 double
@@ -303,5 +303,5 @@ Time::Time(Int64 usec) :
 std::ostream&
 IceUtil::operator<<(std::ostream& out, const Time& tm)
 {
-    return out << tm.toMicroSeconds() / 1000000.0;
+    return out << static_cast<double>(tm.toMicroSeconds()) / 1000000.0;
 }
